moved specMaps to sims_utils
