+++ tests/testPhotometry.py
@@ -8,7 +8,7 @@ from lsst.sims.utils import ObservationMetaData
 from lsst.sims.catalogs.generation.utils import myTestGals, myTestStars, \
                                                 makeStarTestDB, makeGalTestDB, getOneChunk
 
-from lsst.sims.catalogs.measures.instance import defaultSpecMap
+from lsst.sims.utils import defaultSpecMap
 from lsst.sims.photUtils.Bandpass import Bandpass
 from lsst.sims.photUtils.Sed import Sed
 from lsst.sims.photUtils.EBV import EBVbase
